Allow custom cards inside vertical-stack-in-card
- Custom cards in lovelace are not surrounded by the hui-...-card pre- and
suffixes in their html tags. A check is added to setConfig() to check
for custom cards and add them correctly.
- Further, setConfig() of the child cards are called in setConfig() of
vertical-stack-in-card rather than on every update of the hass state.
This is consistent with the behavior of built-in cards and stops child
cards from reinitializing constantly.
- The removal of the setConfig() call from the hass setter allowed for
the removal of the stack_index variable, which means vertical and
horizontal stacks can be placed anywhere inside the
vertical-stack-in-card, and not only as the first child.