Fix creating custom cards inside vertical-stack
when other custom cards are inside the vertical-stack-in-card element. they might not be loaded when homeassistant tries to load the vertical stack.
Especially the "element.setConfig" function crashed when the custom element inside the vertical stack has not been defined.
Using this change, the card checks if the custom card has been defined, and if it is not it will wait until the Promise "customElements.whenDefined" is fullfilled. Then the card is redrawn with all custom elements inside loading properly.
This has been tested with some custom cards, but should work with any type.
The code has been adapted from the lovelace-card-loader by @thomasloven